// frontend/pdm-web/src/client/sdk/forms/multi-selector/config/multi-selector.type.ts
export interface MultiSelectorConfigType {
    key: string;
    initValue: any;
    setItem: any;
    title?: any;
    selectedValue?: any;
    idField?: string;
    labelField?: string;
    isMultiple?: boolean;
    isShowSelectedList?: boolean;
    transformItem?: any;
    invisibleBorder?: boolean;
    showSelectedValue?: boolean;
    isDefaultAllChecked?: boolean;
}